backend/requirement.txt
youtube_transcript_api
google-generativeai
python-dotenv
pathlib
uvicorn
fastapi